receiver/requirements.txt
connexion[flask,swagger-ui,uvicorn]==3.1.0
Flask==3.0.3
Flask-CORS==4.0.0
PyYAML==6.0.1
kafka-python==2.0.2
uvicorn==0.30.1
